requirements.txt
aiofiles==23.2.1
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
attrs==23.2.0
bidict==0.23.1
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
docutils==0.19
exceptiongroup==1.2.0
fastapi==0.109.2
frozenlist==1.4.1
gym==0.26.2
gym-notices==0.0.8
h11==0.14.0
httpcore==1.0.4
httptools==0.6.1
httpx==0.27.0
idna==3.6
ifaddr==0.2.0
importlib_metadata==7.1.0
itsdangerous==2.1.2
Jinja2==3.1.4
langid==1.1.6
ldap3==2.9.1
libsass==0.23.0
markdown2==2.4.13
MarkupSafe==2.1.5
multidict==6.0.5
nes_py==8.2.1
nicegui==1.4.25
niceguiToolkit==0.1.3
numpy==1.26.4
orjson==3.9.15
pandas==2.2.1
profanity==1.1
pscript==0.7.7
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.6.4
pydantic_core==2.16.3
pyglet==1.5.21
Pygments==2.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.0
python-ldap==3.4.4
python-multipart==0.0.9
python-socketio==5.11.2
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.1
starlette==0.36.3
tqdm==4.66.2
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
vbuild==0.8.2
watchfiles==0.21.0
websockets==12.0
wsproto==1.2.0
yarl==1.9.4
zipp==3.18.1
